requirements.txt
requests==2.22.0
beautifulsoup4==4.8.0
progressbar2==3.42.0
